Fix processing fail when cut near start of recording
The comment which said that a negative video offset was ok was in fact
wrong, it was causing all sorts of problems.

Since a cut can't be moved to start at an *earlier* time, the only real
solution is to push cuts to start at a later time instead. A special
case is added for the start of the video where any cuts within the
minimum cut length of the start get pushed back to (exactly) the minimum
cut length.

This can theoretically push multiple cuts back to the same timestamp;
this is not an issue because the following pass will clean that up.