fix: guarantee deafen is executed for dial-in users in lobby
The deafen command is not always executed for dial-in users when
they're in the guest lobby. The centralized deafen method looks for a
valid VoiceUser via its internal user ID to execute deafen, but there
might not be a valid intId registered yet for voice-only users at that
point in time.

Guarantee deafen is always executed for dial-in users when applicable by
splitting deafen into a web user version (looks for valid voice user)
and dial-in version (executes the deafen command directly).